+++ packages/file_selector/file_selector_android/pigeons/file_selector_api.dart
@@ -7,11 +7,10 @@ import 'package:pigeon/pigeon.dart';
 @ConfigurePigeon(
   PigeonOptions(
     dartOut: 'lib/src/file_selector_api.g.dart',
-    javaOut:
-        'android/src/main/java/dev/flutter/packages/file_selector_android/GeneratedFileSelectorApi.java',
-    javaOptions: JavaOptions(
+    kotlinOut:
+        'android/src/main/kotlin/dev/flutter/packages/file_selector_android/GeneratedFileSelectorApi.kt',
+    kotlinOptions: KotlinOptions(
       package: 'dev.flutter.packages.file_selector_android',
-      className: 'GeneratedFileSelectorApi',
     ),
     copyrightHeader: 'pigeons/copyright.txt',
   ),