+++ packages/quick_actions/quick_actions_android/android/src/main/java/io/flutter/plugins/quickactions/QuickActions.java
@@ -11,23 +11,21 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.os.Build;
-import android.os.Handler;
-import android.os.Looper;
 import androidx.annotation.ChecksSdkIntAtLeast;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.core.content.pm.ShortcutInfoCompat;
 import androidx.core.content.pm.ShortcutManagerCompat;
 import androidx.core.graphics.drawable.IconCompat;
-import io.flutter.plugins.quickactions.Messages.AndroidQuickActionsApi;
-import io.flutter.plugins.quickactions.Messages.FlutterError;
-import io.flutter.plugins.quickactions.Messages.ShortcutItemMessage;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.Executor;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import kotlin.Result;
+import kotlin.Unit;
+import kotlin.jvm.functions.Function1;
+import org.jetbrains.annotations.NotNull;
 
 final class QuickActions implements AndroidQuickActionsApi {
   static final String EXTRA_ACTION = "some unique action key";
@@ -48,7 +46,8 @@ final class QuickActions implements AndroidQuickActionsApi {
   }
 
   // Returns true when running on a version of Android that supports quick actions.
-  // When this returns false, methods should silently no-op, per the documented behavior (see README.md).
+  // When this returns false, methods should silently no-op, per the documented behavior (see
+  // README.md).
   @ChecksSdkIntAtLeast(api = Build.VERSION_CODES.N_MR1)
   boolean isVersionAllowed() {
     return Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1;
@@ -56,13 +55,13 @@ final class QuickActions implements AndroidQuickActionsApi {
 
   @Override
   public void setShortcutItems(
-      @NonNull List<ShortcutItemMessage> itemsList, @NonNull Messages.VoidResult result) {
+      @NonNull List<ShortcutItemMessage> itemsList,
+      @NotNull Function1<? super @NotNull Result<@NotNull Unit>, @NotNull Unit> callback) {
     if (!isVersionAllowed()) {
-      result.success();
+      ResultUtilsKt.completeWithUnitSuccess(callback);
       return;
     }
     List<ShortcutInfoCompat> shortcuts = shortcutItemMessageToShortcutInfo(itemsList);
-    Executor uiThreadExecutor = new UiThreadExecutor();
     ThreadPoolExecutor executor =
         new ThreadPoolExecutor(0, 1, 1, TimeUnit.SECONDS, new LinkedBlockingQueue<>());
 
@@ -76,21 +75,16 @@ final class QuickActions implements AndroidQuickActionsApi {
             // Leave dynamicShortcutsSet as false
           }
 
-          final boolean didSucceed = dynamicShortcutsSet;
-
-          // TODO(camsim99): Investigate removing all of the executor logic in favor of background channels.
-          uiThreadExecutor.execute(
-              () -> {
-                if (didSucceed) {
-                  result.success();
-                } else {
-                  result.error(
-                      new FlutterError(
-                          "quick_action_setshortcutitems_failure",
-                          "Exception thrown when setting dynamic shortcuts",
-                          null));
-                }
-              });
+          if (dynamicShortcutsSet) {
+            ResultUtilsKt.completeWithUnitSuccess(callback);
+          } else {
+            ResultUtilsKt.completeWithError(
+                callback,
+                new FlutterError(
+                    "quick_action_setshortcutitems_failure",
+                    "Exception thrown when setting dynamic shortcuts",
+                    null));
+          }
         });
   }
 
@@ -176,14 +170,5 @@ final class QuickActions implements AndroidQuickActionsApi {
         .putExtra(EXTRA_ACTION, type)
         .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
         .addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
-  }
-
-  static class UiThreadExecutor implements Executor {
-    private final Handler handler = new Handler(Looper.getMainLooper());
-
-    @Override
-    public void execute(Runnable command) {
-      handler.post(command);
-    }
   }
 }