+++ packages/video_player/video_player_avfoundation/test/avfoundation_video_player_test.mocks.dart
@@ -7,7 +7,10 @@ import 'dart:async' as _i4;
 
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i3;
-import 'package:video_player_avfoundation/src/messages.g.dart' as _i2;
+import 'package:video_player_avfoundation/src/video_player_instance_messages.g.dart'
+    as _i5;
+import 'package:video_player_avfoundation/src/video_player_plugin_messages.g.dart'
+    as _i2;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -22,6 +25,7 @@ import 'package:video_player_avfoundation/src/messages.g.dart' as _i2;
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
+// ignore_for_file: invalid_use_of_internal_member
 
 class _FakeTexturePlayerIds_0 extends _i1.SmartFake
     implements _i2.TexturePlayerIds {
@@ -111,7 +115,7 @@ class MockAVFoundationVideoPlayerApi extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockVideoPlayerInstanceApi extends _i1.Mock
-    implements _i2.VideoPlayerInstanceApi {
+    implements _i5.VideoPlayerInstanceApi {
   @override
   String get pigeonVar_messageChannelSuffix =>
       (super.noSuchMethod(
@@ -194,6 +198,30 @@ class MockVideoPlayerInstanceApi extends _i1.Mock
   _i4.Future<void> dispose() =>
       (super.noSuchMethod(
             Invocation.method(#dispose, []),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
+
+  @override
+  _i4.Future<List<_i5.MediaSelectionAudioTrackData>> getAudioTracks() =>
+      (super.noSuchMethod(
+            Invocation.method(#getAudioTracks, []),
+            returnValue:
+                _i4.Future<List<_i5.MediaSelectionAudioTrackData>>.value(
+                  <_i5.MediaSelectionAudioTrackData>[],
+                ),
+            returnValueForMissingStub:
+                _i4.Future<List<_i5.MediaSelectionAudioTrackData>>.value(
+                  <_i5.MediaSelectionAudioTrackData>[],
+                ),
+          )
+          as _i4.Future<List<_i5.MediaSelectionAudioTrackData>>);
+
+  @override
+  _i4.Future<void> selectAudioTrack(int? trackIndex) =>
+      (super.noSuchMethod(
+            Invocation.method(#selectAudioTrack, [trackIndex]),
             returnValue: _i4.Future<void>.value(),
             returnValueForMissingStub: _i4.Future<void>.value(),
           )