+++ script/tool/lib/src/analyze_command.dart
@@ -13,6 +13,7 @@ import 'common/gradle.dart';
 import 'common/output_utils.dart';
 import 'common/package_looping_command.dart';
 import 'common/plugin_utils.dart';
+import 'common/pub_utils.dart';
 import 'common/repository_package.dart';
 import 'common/xcode.dart';
 
@@ -357,12 +358,13 @@ class AnalyzeCommand extends PackageLoopingCommand {
   }
 
   Future<bool> _runPubCommand(RepositoryPackage package, String command) async {
-    final int exitCode = await processRunner.runAndStream(
-      flutterCommand,
-      <String>['pub', command],
-      workingDir: package.directory,
+    return runPubCommand(
+      <String>[command],
+      package,
+      processRunner,
+      platform,
+      dartSdkPathOverride: _dartBinaryPath,
     );
-    return exitCode == 0;
   }
 
   /// Runs Gradle lint analysis for the given package, and returns the result
@@ -371,10 +373,11 @@ class AnalyzeCommand extends PackageLoopingCommand {
     RepositoryPackage package,
   ) async {
     if (!pluginSupportsPlatform(
-      platformAndroid,
-      package,
-      requiredMode: PlatformSupport.inline,
-    )) {
+          platformAndroid,
+          package,
+          requiredMode: PlatformSupport.inline,
+        ) ||
+        !package.platformDirectory(FlutterPlatform.android).existsSync()) {
       return PackageResult.skip(
         'Package does not contain native Android plugin code',
       );