+++ packages/file_selector/file_selector_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.2+7
+
+* Migrates to Built-in Kotlin to support AGP 9.
+* Updates minimum supported SDK version to Flutter 3.44/Dart 3.12.
+
+## 0.5.2+6
+
+* Updates internal implementation to use Kotlin Pigeon.
+
 ## 0.5.2+5
 
 * Updates build files from Groovy to Kotlin.