+++ packages/shared_preferences/shared_preferences_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.4.24
+
+* Migrates to Built-in Kotlin to support AGP 9.
+* Updates minimum supported SDK version to Flutter 3.44/Dart 3.12.
+
 ## 2.4.23
 
 * Fixes dartdoc comments that accidentally used HTML.